Move workflow tests out of admin package
Closes #44847
