+++ tests/base/src/test/java/org/keycloak/tests/workflow/BrokeredUserSessionRefreshTimeWorkflowTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.keycloak.tests.admin.model.workflow;
+package org.keycloak.tests.workflow;
 
 import java.time.Duration;
 import java.util.List;
