+++ tests/base/src/test/java/org/keycloak/tests/workflow/GroupMembershipJoinWorkflowTest.java
@@ -1,4 +1,4 @@
-package org.keycloak.tests.admin.model.workflow;
+package org.keycloak.tests.workflow;
 
 import java.time.Duration;
 import java.util.List;
