+++ tests/base/src/test/java/org/keycloak/tests/workflow/WorkflowsScheduledTaskServerConfig.java
@@ -1,4 +1,4 @@
-package org.keycloak.tests.admin.model.workflow;
+package org.keycloak.tests.workflow;
 
 import org.keycloak.models.workflow.WorkflowsEventListenerFactory;
 import org.keycloak.testframework.server.KeycloakServerConfig;
